move status to profile component
